requirements.txt
requests
python-dotenv
numpy
flask
gunicorn
